requirements/base.txt
django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0
django-cors-headers>=4.3,<5.0
django-htmx>=1.17,<2.0
django-celery-beat>=2.6,<3.0
django-storages[s3]>=1.14,<2.0
django-import-export>=4.0,<5.0
django-simple-history>=3.5,<4.0
drf-spectacular>=0.27,<1.0
celery[redis]>=5.4,<6.0
boto3>=1.34,<2.0
Pillow>=10.0,<11.0
gunicorn>=22.0,<23.0
psycopg[binary]>=3.1,<4.0
whitenoise>=6.6,<7.0
python-decouple>=3.8,<4.0
sentry-sdk[django]>=2.0,<3.0
qrcode[pil]>=7.4,<8.0
weasyprint>=62.0,<63.0
